+++ lib/main.dart
@@ -10,9 +10,9 @@ import 'package:little_victories/res/custom_colours.dart';
 import 'package:little_victories/screens/home/debug_screen.dart';
 import 'package:little_victories/screens/home/home_screen.dart';
 import 'package:little_victories/screens/home/preferences_screen.dart';
+import 'package:little_victories/screens/home/view_victories_screen.dart';
 import 'package:little_victories/screens/preferences/push_notifications_screen.dart';
 import 'package:little_victories/screens/sign_in/sign_in_screen.dart';
-import 'package:little_victories/screens/home/view_victories_screen.dart';
 import 'package:little_victories/util/notifications_service.dart';
 import 'package:page_transition/page_transition.dart';
