+++ lib/screens/home/preferences_screen.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:little_victories/data/preferences_model.dart';
-import 'package:little_victories/res/secure_storage.dart';
+import 'package:little_victories/util/secure_storage.dart';
 import 'package:little_victories/widgets/common/custom_button.dart';
 import 'package:little_victories/widgets/common/lv_logo.dart';
 import 'package:little_victories/widgets/common/page_body.dart';
@@ -10,7 +10,7 @@ import 'package:little_victories/widgets/modals/sign_out_of_google_modal.dart';
 import 'package:little_victories/widgets/preferences/reminders_switch_widget.dart';
 import 'package:little_victories/widgets/preferences/reminders_timepicker_widget.dart';
 
-import '../res/constants.dart';
+import '../../res/constants.dart';
 
 class PreferencesScreen extends StatefulWidget {
   const PreferencesScreen({Key? key}) : super(key: key);
