+++ lib/util/notifications_service.dart
@@ -2,7 +2,7 @@ import 'package:awesome_notifications/awesome_notifications.dart';
 import 'package:flutter/material.dart';
 import 'package:little_victories/res/constants.dart';
 import 'package:little_victories/res/custom_colours.dart';
-import 'package:little_victories/res/secure_storage.dart';
+import 'package:little_victories/util/secure_storage.dart';
 import 'package:little_victories/widgets/modals/notifications_consent_modal.dart';
 
 const String kNotificationChannelKeyReminders =
