+++ lib/widgets/modals/notifications_consent_modal.dart
@@ -2,8 +2,8 @@ import 'package:awesome_notifications/awesome_notifications.dart';
 import 'package:flutter/material.dart';
 import 'package:little_victories/res/constants.dart';
 import 'package:little_victories/res/custom_colours.dart';
-import 'package:little_victories/res/secure_storage.dart';
 import 'package:little_victories/util/notifications_service.dart';
+import 'package:little_victories/util/secure_storage.dart';
 import 'package:little_victories/util/utils.dart';
 
 class NotificationsConsentModal extends StatefulWidget {
