+++ lib/widgets/preferences/reminders_switch_widget.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:little_victories/res/constants.dart';
 import 'package:little_victories/res/custom_colours.dart';
-import 'package:little_victories/res/secure_storage.dart';
 import 'package:little_victories/util/notifications_service.dart';
+import 'package:little_victories/util/secure_storage.dart';
 
 class RemindersSwitchWidget extends StatefulWidget {
   const RemindersSwitchWidget({
