+++ lib/widgets/preferences/reminders_timepicker_widget.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:little_victories/res/constants.dart';
 import 'package:little_victories/res/custom_colours.dart';
-import 'package:little_victories/res/secure_storage.dart';
 import 'package:little_victories/util/notifications_service.dart';
+import 'package:little_victories/util/secure_storage.dart';
 
 class ReminderTimepickerWidget extends StatefulWidget {
   const ReminderTimepickerWidget({
