+++ pubspec.yaml
@@ -45,6 +45,7 @@ dependencies:
   # Widget Packages
   expansion_tile_card: ^2.0.0
   confetti: ^0.7.0
+  intro_views_flutter: ^3.2.0
 
   # Other Packages
   flutter_launcher_icons: ^0.9.2
